// libs/features/complaints/src/lib/shared/enums/itsm-type-subtype.ts
export enum TypeSubtypeArEnum {
    Ameen='أمين',
    AmeenEstablishments ='منشآت',
    AmeenContributors = 'المشتركين',
    AmeenVoluntaryContributors = 'مشترك اختياري',
    AmeenContributionCollection = 'الاشتراكات والتحصيل',
    AmeenOH = 'الأخطار المهنية',
    AmeenCRM = 'العناية بالعملاء',
    AmeenCore = 'الخدمات الأساسية',
    AmeenViolations = ' المخالفات',
    AmeenBenefits = 'أمين منافع',
    Establishments = 'منشآت',
    Contract = 'المشتركين',
    Contributors = 'الاشتراكات والتحصيل',
    OccupationHazards = 'الأخطار المهنية',
    CRM = 'العناية بالعملاء',
    CRMService = 'العناية بالعملاء',
    Coreservices = 'الخدمات الأساسية',
    Violations = 'المخالفات',
    BeneficiariesSupport = 'المنافع',
    Heirssupport = 'الورثة',
    ProActiveEstRegistration = 'تسجيل الإستباقي للمنشآت',
    EstablishmentInquiry = 'الاستعلام عن المنشأة',
    EstablishmentRegistration = ' تسجيل المنشأة',
    DisplayEstablishmentData = 'عرض بيانات المنشأة',
    UpdateEstablishmentData = 'تحديث بيانات المنشأة',
    GOSICertificates = ' الشهادات',
    EstablishmentTermination = 'إنهاء نشاط المنشأة',
    SupervisorAuthority = 'صلاحيات المشرفين',
    EstablishmentRelation = '(ربط/فصل) منشأة',
    ModifyEstablishmentHQ = 'عديل المركز الرئيسي للمنشأة',
    ChangeEstablishmentOwner = 'تغيير ملكية منشأة',
    ModifyContributorPeriod = 'تعديل مدد مشترك',
    ModifyContributorsWages = 'تحديث أجور المشتركين',
    DisplayContributorsData = 'عرض بيانات المشتركين',
    ContributorsRegistrationOrRemoval = 'تسجيل/استبعاد المشتركين',
    ContractsContributorsAgreements = 'توثيق عقود المشتركين',
    UpdateContributorsDatawithHRSD = ' تحديث بيانات المشتركين مع وزارة الموارد البشرية',
    RelocateContributors = 'نقل المشتركين',
    VoluntaryContributorRegistrationOrRemoval = ' تسجيل/استبعاد المشترك الاختياري',
    ModifyVoluntaryContributorPeriod = 'تعديل مدد مشترك اختياري',
    DisplayVoluntaryContributorData = ' عرض بيانات مشترك اختياري',
    ModifyVoluntaryContributorWages = 'تحديث أجور مشترك اختياري',
    BillsAndReceipts = ' فواتير وايصالات',
    Notifications = 'الاشعارات',
    SADADPayments = 'سداد فواتير',
    ManagingAccountsReceivable = 'دارة الأرصدة الدائنة',
    Bills = 'فواتير',
    Installment = 'تقسيط',
    Exemptionfromfines = 'الإعفاء من الغرامات',
    OHClaims = 'مطالبات',
    PaymentRefund = 'استعادة المدفوعات',
    injuries = 'الإصابات',
    Hospitals = 'المستشفيات',
    Treatment = 'علاج',
    CustomerPersonalDataManagement = 'إدارة البيانات الشخصية',
    Documents = 'المستندات',
    Transactions = 'المعاملات',
    Login = 'تسجيل الدخول',
    TextMessages = 'الرسائل النصية',
    Permissions = ' الصلاحيات',
    AddingViolation = 'تسجيل مخالفة',
    EditingViolation = 'تعديل مخالفة',
    CancellationofViolation = ' الغاء مخالفة',
    SANED = 'ساند',
    Occupational = 'عجز مهني',
    NonOccupational = 'عجز غير مهني',
    RetirementLumpsumBenefit = 'دفعة تقاعد',
    RetirementPensionBenefit = 'معاش تقاعد',
    HeirPensionBenefit = 'معاش ورثة',
    HeirLumpsumBenefit = 'دفعة ورثة',
    
    Closed = 'إنتهاء النشاط',
    Resolved = 'تم الإنتهاء من حل بلاغ الدعم الفني',
    Reopened = 'إعادة فتح',
    InProgress = 'قيد المعالجة',
    Assigned = 'مسندة',

    Complaint = "شكوى",
    Enquiry = "استفسار",
    Suggestion = "اقتراح",
    Appeal = "اعتراض",
    Plea = "التماس",

}